feature: independent profile name component
